python_data_analytics/07_Desenvolvimento Web/FastApi/dio_blog/src/schemas/auth_schema.py
from pydantic import BaseModel


class AccessToken(BaseModel):
    iss: str
    sub: int
    aud: str
    exp: float
    iat: float
    nbf: float
    jti: str


class JwtToken(BaseModel):
    access_token: AccessToken


class LoginIn(BaseModel):
    user_id: str
