python_data_analytics/07_Desenvolvimento Web/FastApi/dio_blog/pyproject.toml
[tool.poetry]
name = "dio-blog"
version = "0.1.0"
description = "project to create and learn FastAPI"
authors = ["Gabryel_Brz <gabrielbarbosa.alternativa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
databases = {extras = ["aiosqlite"], version = "^0.9.0"}
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
pytest-mock = "^3.14.0"
httpx = "^0.28.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
quote-style = "single"
